input_jetstream: allow using metadata to configure nak_delay per-message
This uses meta("nak_delay") to change the behavior of input jetstream
when an error occurs.

It seems like a bit of a hack, as no other inputs use MetaGet. Plausible
there's a better way to flow a dynamic value here, perhaps by defining
another blobl program to compute it when needed.